+++ src/TapkuLibrary/TKCalendarMonthView.m
@@ -529,7 +529,7 @@ static UIImage *tileImage;
 	}
 	
 	[self addSubview:self.selectedImageView];
-	self.currentDay.text = [NSString stringWithFormat:@"%d",day];
+	self.currentDay.text = [NSString stringWithFormat:@"%ld",(long)day];
 	
 	if (self.marks.count > 0) {
 		if([self.marks[row * 7 + column] boolValue])
