+++ src/TapkuLibrary/UIImage+TKCategory.m
@@ -85,7 +85,7 @@
 - (UIImage *) imageByApplyingTintEffectWithColor:(UIColor *)tintColor{
     const CGFloat EffectColorAlpha = 0.6;
     UIColor *effectColor = tintColor;
-    int componentCount = CGColorGetNumberOfComponents(tintColor.CGColor);
+    NSInteger componentCount = CGColorGetNumberOfComponents(tintColor.CGColor);
     if (componentCount == 2) {
         CGFloat b;
         if ([tintColor getWhite:&b alpha:NULL]) {
